refactor: remove unreachable Mac UI island and post-unification debris
