--- Apps/Mac/Peekaboo/Core/AgentEventStream.swift
@@ -1,6 +1,0 @@
-import Foundation
-import PeekabooCore
-
-// Re-export types from PeekabooCore
-public typealias AgentEvent = PeekabooCore.AgentEvent
-public typealias AgentEventDelegate = PeekabooCore.AgentEventDelegate
