+++ Apps/Mac/Peekaboo/Core/ToolFormatterBridge.swift
@@ -14,10 +14,10 @@ class ToolFormatterBridge {
     private init() {}
 
     /// Format tool call for display in the Mac app
-    func formatToolCall(name: String, arguments: String, result: String? = nil) -> String {
+    func formatToolCall(name: String, arguments: String) -> String {
         // Parse tool type
         guard let toolType = ToolType(rawValue: name) else {
-            return self.formatUnknownTool(name: name, arguments: arguments, result: result)
+            return self.formatUnknownTool(name: name)
         }
 
         // Get formatter from registry
@@ -26,30 +26,11 @@ class ToolFormatterBridge {
         // Parse arguments
         let args = self.parseArguments(arguments)
 
-        if let result {
-            // Format completed tool call
-            let resultDict = self.parseArguments(result)
-            let success = (resultDict["success"] as? Bool) ?? true
-            let summaryText = ToolEventSummary.from(resultJSON: resultDict)?
-                .shortDescription(toolName: name) ?? formatter.formatResultSummary(result: resultDict)
-
-            if success {
-                if !summaryText.isEmpty {
-                    return "\(AgentDisplayTokens.Status.success) \(toolType.displayName): \(summaryText)"
-                }
-                return "\(AgentDisplayTokens.Status.success) \(toolType.displayName) completed"
-            } else {
-                let error = (resultDict["error"] as? String) ?? "Failed"
-                return "\(AgentDisplayTokens.Status.failure) \(toolType.displayName): \(error)"
-            }
+        let summary = formatter.formatCompactSummary(arguments: args)
+        if !summary.isEmpty {
+            return "\(AgentDisplayTokens.Status.running) \(toolType.displayName): \(summary)"
         } else {
-            // Format tool call in progress
-            let summary = formatter.formatCompactSummary(arguments: args)
-            if !summary.isEmpty {
-                return "\(AgentDisplayTokens.Status.running) \(toolType.displayName): \(summary)"
-            } else {
-                return "\(AgentDisplayTokens.Status.running) \(toolType.displayName)"
-            }
+            return "\(AgentDisplayTokens.Status.running) \(toolType.displayName)"
         }
     }
 
@@ -69,29 +50,6 @@ class ToolFormatterBridge {
 
         // Fall back to formatted JSON
         return self.formatJSON(arguments)
-    }
-
-    /// Format tool result for detailed view
-    func formatResult(name: String, result: String) -> String {
-        guard let toolType = ToolType(rawValue: name) else {
-            return result
-        }
-
-        let formatter = ToolFormatterRegistry.shared.formatter(for: toolType)
-        let resultDict = self.parseArguments(result)
-        if let summary = ToolEventSummary.from(resultJSON: resultDict)?.shortDescription(toolName: name),
-           !summary.isEmpty
-        {
-            return summary
-        }
-
-        let summary = formatter.formatResultSummary(result: resultDict)
-        if !summary.isEmpty {
-            return summary
-        }
-
-        // Fall back to formatted JSON
-        return self.formatJSON(result)
     }
 
     /// Get icon for tool
@@ -134,34 +92,8 @@ class ToolFormatterBridge {
         return result
     }
 
-    private func formatUnknownTool(name: String, arguments: String, result: String?) -> String {
+    private func formatUnknownTool(name: String) -> String {
         let displayName = self.toolDisplayName(for: name)
-
-        if let result {
-            let resultDict = self.parseArguments(result)
-            let success = (resultDict["success"] as? Bool) ?? true
-            let summaryText = ToolEventSummary.from(resultJSON: resultDict)?.shortDescription(toolName: name)
-
-            if success {
-                if let summaryText, !summaryText.isEmpty {
-                    return "\(AgentDisplayTokens.Status.success) \(displayName): \(summaryText)"
-                }
-                return "\(AgentDisplayTokens.Status.success) \(displayName) completed"
-            } else {
-                let error = (resultDict["error"] as? String) ?? "Failed"
-                return "\(AgentDisplayTokens.Status.failure) \(displayName): \(error)"
-            }
-        } else {
-            return "\(AgentDisplayTokens.Status.running) \(displayName)"
-        }
-    }
-}
-
-// MARK: - ToolType Extension for Mac App
-
-extension ToolType {
-    /// Icon to use in Mac app UI
-    var icon: String {
-        AgentDisplayTokens.icon(for: self.rawValue)
+        return "\(AgentDisplayTokens.Status.running) \(displayName)"
     }
 }
