+++ Apps/Mac/PeekabooTests/Services/SettingsServiceTests.swift
@@ -833,15 +833,6 @@ struct PeekabooSettingsConfigHydrationTests {
             "claude-fable-5",
             "claude-sonnet-5",
         ])
-        #expect(AIAssistantModelCatalog.options.map(\.model.modelId) == [
-            "gpt-5.6-sol",
-            "gpt-5.6-terra",
-            "gpt-5.6-luna",
-            "gpt-5.5",
-            "claude-fable-5",
-            "claude-sonnet-5",
-            "claude-opus-4-8",
-        ])
     }
 
     @Test
