+++ Core/PeekabooCore/Sources/PeekabooAgentRuntime/ToolFormatting/ToolResultExtractor.swift
@@ -89,12 +89,39 @@ public enum ToolResultExtractor {
         return nil
     }
 
-    // MARK: - Double Extraction (legacy helper)
+    // MARK: - Double Extraction
 
-    /// Extract a Double value from the result (legacy; prefer the unified method below)
+    /// Extract a Double value from the result
     public static func double(_ key: String, from result: [String: Any]) -> Double? {
-        // Delegate to unified implementation below
-        self.doubleUnified(key, from: result)
+        if let value = result[key] as? Double {
+            return value
+        }
+        if let value = result[key] as? Int {
+            return Double(value)
+        }
+        if let stringValue = string(key, from: result), let d = Double(stringValue) {
+            return d
+        }
+        if let wrapper = result[key] as? [String: Any] {
+            if let v = wrapper["value"] as? Double {
+                return v
+            }
+            if let v = wrapper["value"] as? Int {
+                return Double(v)
+            }
+            if let v = wrapper["value"] as? String, let d = Double(v) {
+                return d
+            }
+        }
+        if let data = result["data"] as? [String: Any] {
+            if let v = data[key] as? Double {
+                return v
+            }
+            if let v = data[key] as? Int {
+                return Double(v)
+            }
+        }
+        return nil
     }
 
     // MARK: - Boolean Extraction
@@ -132,42 +159,6 @@ public enum ToolResultExtractor {
             return value
         }
 
-        return nil
-    }
-
-    // MARK: - Number Extraction
-
-    /// Extract a Double value from the result (unified)
-    public static func doubleUnified(_ key: String, from result: [String: Any]) -> Double? {
-        // Extract a Double value from the result (unified)
-        if let value = result[key] as? Double {
-            return value
-        }
-        if let value = result[key] as? Int {
-            return Double(value)
-        }
-        if let stringValue = string(key, from: result), let d = Double(stringValue) {
-            return d
-        }
-        if let wrapper = result[key] as? [String: Any] {
-            if let v = wrapper["value"] as? Double {
-                return v
-            }
-            if let v = wrapper["value"] as? Int {
-                return Double(v)
-            }
-            if let v = wrapper["value"] as? String, let d = Double(v) {
-                return d
-            }
-        }
-        if let data = result["data"] as? [String: Any] {
-            if let v = data[key] as? Double {
-                return v
-            }
-            if let v = data[key] as? Int {
-                return Double(v)
-            }
-        }
         return nil
     }
 
@@ -302,20 +293,5 @@ public enum ToolResultExtractor {
 
         // Default to true if no explicit failure indicators
         return true
-    }
-
-    // MARK: - Unwrapping Utilities
-
-    /// Unwrap a potentially nested result structure
-    public static func unwrapResult(_ result: [String: Any]) -> [String: Any] {
-        // Check for wrapped format {"type": "object", "value": {...}}
-        if result["type"] as? String == "object",
-           let value = result["value"] as? [String: Any]
-        {
-            return value
-        }
-
-        // Return as-is if not wrapped
-        return result
     }
 }
